[ADD] Clicks statistics in mass_mailings
