+++ addons/mass_mailing/models/mass_mailing.py
@@ -493,7 +493,7 @@ class MassMailing(osv.Model):
         'email_from': fields.char('From', required=True),
         'create_date': fields.datetime('Creation Date'),
         'sent_date': fields.datetime('Sent Date', oldname='date', copy=False),
-        'schedule_date': fields.datetime('Scheduled Send Date'),
+        'schedule_date': fields.datetime('Schedule in the Future'),
         'body_html': fields.html('Body'),
         'attachment_ids': fields.many2many(
             'ir.attachment', 'mass_mailing_ir_attachments_rel',
@@ -790,12 +790,43 @@ class MassMailing(osv.Model):
             self.write(cr, uid, [mailing.id], {'sent_date': fields.datetime.now(), 'state': 'done'}, context=context)
         return True
 
+    # List of link's href that will not be converted by the shortener
+    links_backlist = ['/unsubscribe_from_list']
+
     def convert_links(self, cr, uid, ids, context=None):
         res = {}
         for mass_mailing in self.browse(cr, uid, ids, context=context):
-            res[mass_mailing.id] = mass_mailing.body_html if mass_mailing.body_html else ''
             utm_mixin = mass_mailing.mass_mailing_campaign_id if mass_mailing.mass_mailing_campaign_id else mass_mailing
-            res[mass_mailing.id] = self.pool['website.links'].convert_links(cr, uid, res[mass_mailing.id], utm_mixin, context=context)            
+            html = mass_mailing.body_html if mass_mailing.body_html else ''
+
+            for match in re.findall(URL_REGEX, html):
+                href = match[0]
+                long_url = match[1]
+
+                if not [s for s in self.links_backlist if s in long_url]:
+                    vals = {
+                        'url': long_url,
+                        'mass_mailing_id': mass_mailing.id
+                    }
+
+                    if mass_mailing.mass_mailing_campaign_id:
+                        vals['mass_mailing_campaign_id'] = mass_mailing.mass_mailing_campaign_id.id
+
+                    if utm_mixin.campaign_id:
+                        vals['campaign_id'] = utm_mixin.campaign_id.id
+                    if utm_mixin.source_id:
+                        vals['source_id'] = utm_mixin.source_id.id
+                    if utm_mixin.medium_id:
+                        vals['medium_id'] = utm_mixin.medium_id.id
+
+                    link = self.pool.get('website.links').create(cr, uid, vals, context)
+                    shorten_url = self.pool.get('website.links').browse(cr, uid, link, context)[0].short_url
+
+                    if shorten_url:
+                        new_href = href.replace(long_url, shorten_url)
+                        html = html.replace(href, new_href)
+
+            res[mass_mailing.id] = html
         return res
 
     def put_in_queue(self, cr, uid, ids, context=None):
