+++ addons/mass_mailing/static/src/css/mass_mailing.css
@@ -4,7 +4,7 @@
 }
 .openerp .oe_kanban_view .oe_kanban_mass_mailing {
     width: 280px;
-    min-height: 141px;
+    min-height: 35px;
 }
 .openerp .oe_margin_top_8 {
 	margin-top: 8px;
